ml-service/requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
langchain==0.3.7
langchain-community==0.3.7
langchain-openai==0.2.9
langchain-chroma==0.1.4
langchain-core==0.3.19
langchain-text-splitters==0.3.2
chromadb==0.5.23
openai==1.55.3
tiktoken==0.8.0
